Split rejection/exception handlers, clarify http app flow
